Add meSCV::grad_op() - kokkos style; with unit tests (#251)
Design note: old style non-kokkos views methods not implemented.

* support new ElemData/Scratch interface to SCV_GRAD_OP.
This is required for higher-order turbulence source terms
and a variety of moving mesh algorithms that require SCV::grad_op().

* master element rework... Some free functions needed to be moved
up in the file for the SCV element to use them. Formerly, these free
functions were hidden with SCS methods. Also, a few SCS member functions
needed to move to free functions for sharing between SCS and SCV..

* Add tri3d to NGP-ready logic for consolidated kernel usage; this was missinfg before.

* Add unit test concept for dndx_scv; some plumbing work here..
The design coded was intended to compare a new and old ME implementation,
not the correctness. As such, modified tests to compare against a new gold
standard. This requires some if then logic. However, given we only have a scv::grad_op,
it seems managable for now.

Notes:

a) Tested in my other branch on changing mesh motion algorithm.