Alter face-element algorithm looping to ensure uniform face-element ordinals in simd groups.
Our master-element and kernel implementations require that within
a simd group, all faces have the same element-ordinal.

Created a unit-test which alters a mesh to ensure at least two faces
with different element-ordinals are in the same bucket, then added
a custom kernel::execute method which checks for different ordinals
in the same simd group. Then put the corrected algorithm looping
in place to make the test pass.