Created ElemDataRequestsNGP class.
The purpose of this class is to copy the contents of
ElemDataRequests to a GPU device for use by ScratchViews etc.

Data-enums and fields are stored in kokkos views instead of
std sets and maps.

Unit test in UnitTestElemDataRequests.C shows how this class
could be used to copy the contents of a regular ElemDataRequests
class to a gpu device for access within a parallel_for loop.

Still need to address the master-element pointers, which need
to be allocated on device in order to be used on device...