Fix memory leak in HypreLinearSystem
A memory leak occured in HYPRE's hypre_AuxParCSRMatrixInitialize
function when invoked from HypreLinearSystem. This was caused because
HYPRE_IJMatrixInitialized was invoked twice without an intervening call
to HYPRE_IJMatrixAssemble, causing data structures
hypre_AuxParCSRMatrixAuxJ and hypre_AuxParCSRMatrixAuxData to be
reallocated without deallocating the current memory.

The situation occurs at the first outer iteration of an Equation System
(in the first step for static mesh applications and every timestep for
 dynamic mesh applications). To fix this issue a boolean flag is
introduced to call IJMatrixInitialize (in zeroSystem) only if an
IJMatrixAssemble call was made since finalizeLinearSystem was invoked.

This fix was checked using valgrind on debug build.