Wall distance calculation equation system
Solve a Poisson equation to compute the wall-distance for use with k-omega SST
and SST-DES turbulence models. The wall-distance is computed once if there is no
mesh-motion and every timestep when there is mesh-motion. The user has an option
to control the frequency of wall-distance recomputation by setting the
update_frequency parameter such that the distance is recomputed after a certain
number of timesteps.

Updated docs to describe the equation solved and add reference.

Update airfoilRANSEdge case to use the distance calculator instead of the NDTW
field in the exodus input file.