ScratchViews prototype for execution on GPUs (#63)
This pull request implements a unit test that demonstrates the use of ngp::Mesh, ngp::Field in constructing and executing a TestKernel with ScratchViews on GPU. It introduces a new class ElemDataRequestsGPU to ease the transition of host-only kernels to NGP kernels. The *GPU classes will be merged with the original classes ScratchViews and ElemDataRequests once the transition of the core Kernel classes are complete.