Added total-num-fields to ElemDataRequestsNGP.
This is a small pre-requisite step towards bringing the
NGP classes ScratchViewsNGP and ElemDataRequestsGPU into
the mainstream.

Also cleaned up a minor error in KokkosInterface.h related
to 4D and 5D scratch-views. Kokkos doesn't let us use views
of that dimension, but still seemed reasonable to correct
the error in our API.