Finish off this master element GPU unit test.
This converts a couple of functions in the Hex8 and Hex27
master element classes to work on the GPU. I notice that
the Hex27 used some Kokkos views instead of static length
data arrays. Both are valid for the GPU but the Views
should be able to take the place of std::vector for
variable sized data arrays.