Fix a bug from my 'fix intel segfault' commit.
I wasn't allowing for how big 'maxOrdinal' could be
in MultiDimViews, and didn't have a runtime check on it.
Now I'm storing the ordinals in a view sized at runtime
so there's no need for maxOrdinal.
This was a clear array-bounds issue, I don't know how it
got through before...