Overset: Fix selector logic when using optimized ghosting
Use consistent node/element selectors so that the nodes looped over do not
change with mesh motion when using the optimized ghosting introduced in b5a158f.
Previously, ghosting was destroyed prior to recomputing holecut with TIOGA,
with the new ghosting, this does not happen so the parts must exclude ghosted
nodes/elements.

Communicate coords field for overset with TIOGA