Fix for compiler warnings and unit test failure.
The unit test failure was a read past the end of an std::array
as returned from gauss_legendre_rule introduced when the Edge
master element was changed from using std::vector to fixed length
arrays.

This is going to have to eventually be fixed so the
gauss_legendre_rule function can be called on the
GPU without returning std::vector.