NGP looping abstractions (#287)
Introduce wrapper functions that accept a C++ lambda or functor to abstract out
the looping process over buckets, and the handling of ScratchViews for element,
face/element based looping constructs. These constructs will be used to replace
current algorithms that deal with field modification so that the can occur on
device instead of host.

- NgpTypes: Introduce data structures to ease handling NGP data. In preparation
for replacing logic within the code that currently use MetaData/Bulkdata with
ngp::Mesh and ngp::Field<T>, start introducing some convenience types. Replace
handling of ngp data in Realm with the one using wrapper.

- Check looping strategy and field modification using mass flow rate computation
in unit tests.

- Convert massFlowRate calculator in unit tests to use NGP loops