Fix for NALU_ALIGN for Summit. (#315)
* Fix for NALU_ALIGN for Summit.

We were using KOKKOS_MEMORY_ALIGNMENT (which is 64) to align
static arrays.  On Summit this gives a warning that it
is greater than 16 and it appears that because of the warning
the alignment attribute is ignored.

This commmit will change to alignof(std::max_align_t)
if compiling for the GPU.  Not sure this will compile
everywhere.

I seem to recall that KOKKOS_MEMORY_ALIGNMENT was set to 64
because this was needed for SIMD and I think Nalu even
pushed for this a couple of years ago. So I'm not sure the
Kokkos macro can be changed.

* More conservative formulation, from a C++ language standard point of view.

Fixes #314